+++ R/my_first_fcn.R
@@ -1,0 +1,21 @@
+#' Onboarding dummy function
+#'
+#'@author Jian Wang
+#'
+#'@description
+#'This function displays a welcome message and return a string literal to the caller
+#'#'
+#'@export
+#'
+#'@return A character string with a welcome message
+#'
+#'@examples
+#'my_first_fcn()
+#'
+my_first_fcn <- function() {
+
+  message("Welcome to the admiral family!")
+
+  "Welcome to the admiral family!"
+
+}